vision/requirements.txt
absl-py
tensorflow-gpu==1.15.2
tensorflow-datasets==3.1.0
tensorflow-hub==0.8.0
